# backend/internal/docs/swagger.yaml
basePath: /
definitions:
  github_com_studentinovisad_popisomator_backend_internal_dto.AddUpdateItemPropertyRequest:
    properties:
      item_id:
        type: integer
      property_id:
        type: integer
      value:
        type: string
    required:
    - item_id
    - property_id
    - value
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.CreateItemRequest:
    properties:
      amount:
        maximum: 100
        type: integer
      properties:
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemProperty'
        type: array
      type_id:
        type: integer
    required:
    - amount
    - type_id
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.Item:
    properties:
      consumption:
        $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_repository.ConsumptionStatus'
      derived_name:
        type: string
      id:
        type: integer
      properties:
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemProperty'
        type: array
      type_id:
        type: integer
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.ItemProperty:
    properties:
      id:
        type: integer
      value:
        type: string
      visibility:
        type: string
    required:
    - id
    - value
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.ItemsPage:
    properties:
      items:
        description: Item page items
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Item'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.UpdateItemRequest:
    properties:
      consumption:
        enum:
        - not_consumed
        - partially_consumed
        - fully_consumed
        - damaged
        type: string
      id:
        type: integer
      type_id:
        type: integer
    required:
    - id
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_repository.ConsumptionStatus:
    enum:
    - not_consumed
    - partially_consumed
    - fully_consumed
    - damaged
    type: string
    x-enum-varnames:
    - ConsumptionStatusNotConsumed
    - ConsumptionStatusPartiallyConsumed
    - ConsumptionStatusFullyConsumed
    - ConsumptionStatusDamaged
...
  github_com_studentinovisad_popisomator_backend_internal_response.Error:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Inventory management API for Popisomator.
  title: Popisomator API
  version: "1.0"
paths:
...
  /items:
    get:
      parameters:
      - description: Page size (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by item type ID
        in: query
        name: type_id
        type: integer
      - collectionFormat: csv
        description: Filter by consumption status (comma-separated)
        in: query
        items:
          enum:
          - not_consumed
          - partially_consumed
          - fully_consumed
          - damaged
          type: string
        name: consumption
        type: array
      - description: Filter by creation time, RFC3339
        in: query
        name: created_from
        type: string
      - description: Filter by creation time, RFC3339
        in: query
        name: created_to
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemsPage'
        "400":
          description: invalid query parameters
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: List items
      tags:
      - Items
    post:
      consumes:
      - application/json
      parameters:
      - description: Item(s) to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.CreateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Item'
            type: array
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Create one or more items of a type (manager/admin only)
      tags:
      - Items
  /items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: invalid item id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Delete an item (manager/admin only)
      tags:
      - Items
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Item'
        "400":
          description: invalid item id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Get an item by ID
      tags:
      - Items
    patch:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Item'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update an item's type (manager/admin only)
      tags:
      - Items
  /items/{id}/consume:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Consumption status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Item'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update an item's consumption status
      tags:
      - Items
  /items/{id}/properties:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.AddUpdateItemPropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemProperty'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Add a property value to an item (manager/admin only)
      tags:
      - Items
  /items/{id}/properties/{prop_id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property ID
        in: path
        name: prop_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: invalid item/property id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Remove a property value from an item (manager/admin only)
      tags:
      - Items
    put:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property ID
        in: path
        name: prop_id
        required: true
        type: integer
      - description: Property value
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.AddUpdateItemPropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemProperty'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update a property value on an item (manager/admin only)
      tags:
      - Items
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"